requirements.txt
aiodns                        3.2.0
aiofile                       3.9.0
aiohappyeyeballs              2.6.1
aiohttp                       3.11.16
aiosignal                     1.3.2
asgiref                       3.8.1
async-timeout                 5.0.1
attrs                         25.3.0
CacheControl                  0.14.2
cachetools                    5.5.2
caio                          0.9.21
certifi                       2025.1.31
cffi                          1.17.1
charset-normalizer            3.4.1
cryptography                  44.0.2
Django                        5.1.4
django-cors-headers           4.7.0
django-filter                 25.1
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
drf-yasg                      1.21.10
firebase-admin                6.7.0
frozenlist                    1.5.0
google-api-core               2.24.2
google-api-python-client      2.166.0
google-auth                   2.38.0
google-auth-httplib2          0.2.0
google-cloud-core             2.4.3
google-cloud-firestore        2.20.1
google-cloud-storage          3.1.0
google-crc32c                 1.7.1
google-resumable-media        2.7.2
googleapis-common-protos      1.69.2
grpcio                        1.71.0
grpcio-status                 1.71.0
httplib2                      0.22.0
idna                          3.10
inflection                    0.5.1
msgpack                       1.1.0
multidict                     6.4.3
packaging                     24.2
pillow                        11.1.0
pip                           24.3.1
propcache                     0.3.1
proto-plus                    1.26.1
protobuf                      5.29.4
pyasn1                        0.6.1
pyasn1_modules                0.4.2
pycares                       4.6.0
pycparser                     2.22
PyJWT                         2.10.1
pyparsing                     3.2.3
python-dotenv                 1.1.0
pytz                          2025.2
PyYAML                        6.0.2
requests                      2.32.3
rsa                           4.9
setuptools                    75.8.0
sqlparse                      0.5.3
stream-chat                   4.24.0
typing_extensions             4.13.1
uritemplate                   4.1.1
urllib3                       2.3.0
wheel                         0.45.1
yarl                          1.19.0